+++ app/src/main/java/helium314/keyboard/latin/settings/Settings.java
@@ -214,6 +214,7 @@ public final class Settings implements SharedPreferences.OnSharedPreferenceChang
 
     public static final String PREF_SPACE_TO_CHANGE_LANG = "prefs_long_press_keyboard_to_change_lang";
     public static final String PREF_LANGUAGE_SWIPE_DISTANCE = "language_swipe_distance";
+    public static final String PREF_SOURCE_KEY_ACTION_TARGETS = "source_key_action_targets";
     public static final String PREF_TOUCHPAD_SENSITIVITY = "touchpad_sensitivity";
     public static final String PREF_TOUCHPAD_EDGE_SCROLL = "touchpad_edge_scroll";
     public static final String PREF_PERSIST_FLOATING_KEYBOARD = "persist_floating_keyboard";